fix: add_object

+++ src/parse_file/utils/free_splitted.c
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/19 21:36:01 by kjikuhar          #+#    #+#             */
-/*   Updated: 2026/06/12 00:40:33 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/06/12 14:15:25 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	free_splitted(char **splitted)
 {
-	int	i;
+	size_t	i;
 
 	if (splitted == NULL)
 		return ;
